src/source/_apps/elara/pyproject.toml
[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "elara"
version = "0.1.0"
description = "Plumb the depths of generative AI."
readme = "README.md"
authors = [{name = "Grant Moore"}]
license = {text = "MIT"}
requires-python = ">=3.8"

dependencies = [
    "google-generativeai==0.8.3",
    "Jinja2==3.1.5",
    "requests==2.25.1"
]

[project.optional-dependencies]
dev = [
    "pytest"
]

[project.scripts]
elara = "elara.main:main"